Added deletion of matplotlib plots, and exception.
Exception is for Sakarin Villapaitapeli when it crashes by not finding
the files. I'm unsure of why exactly this happens, but it might be some
sort of timeout thing for reading the file too slowly? Now the function
will call itself again if it fails to find the file in the first round.
This needs to be tested in practise.